requirements.txt
# Core DS stack
pandas==2.2.0
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
scipy==1.12.0

# Machine Learning
scikit-learn==1.4.1.post1
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.3
category_encoders==2.6.3
shap==0.44.1

# Time Series
prophet==1.1.5
statsmodels==0.14.1

# Engineering & Utilities
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.66.2
python-dotenv==1.0.1
mlflow==2.11.1
streamlit==1.32.0
fpdf2==2.7.8
evidently==0.4.18
pytest==8.0.2

# Notebooks
jupyterlab==4.1.2
ipykernel==6.29.3
